requirements.txt
python-telegram-bot==20.8
Pillow==10.3.0